+++ guillotina/interfaces/content.py
@@ -8,6 +8,7 @@ from typing import Optional
 from typing import TYPE_CHECKING
 from zope.interface import Attribute
 from zope.interface import Interface
+from zope.interface.common.mapping import IEnumerableMapping
 
 import guillotina.schema
 
@@ -245,7 +246,7 @@ class IItem(IResource):
     """
 
 
-class IAnnotationData(Interface):
+class IAnnotationData(IBaseObject, IEnumerableMapping):
     """
     """