+++ CHANGELOG.rst
@@ -4,6 +4,12 @@ CHANGELOG
 5.3.36 (unreleased)
 -------------------
 
+- Provide patch operations for json field
+  [vangheem]
+
+- Optimize extend operation for bucket list field
+  [vangheem]
+
 - `.` and `..` should be blocked as valid ids. The browser will auto translate them
   to what current dir and parent dir respectively which gives unexpected results.
   [vangheem]